% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_matches.R
\name{summarize_matches}
\alias{summarize_matches}
\title{Summarize primer-template matches}
\usage{
summarize_matches(
  pattern,
  subject_filepath,
  subject_format,
  max.mismatch,
  with.indels = TRUE
)
}
\arguments{
\item{pattern}{A string consisting of the pattern (primer sequence) to look for.
IUPAC ambiguity codes are supported (e.g. "R" matches "A" and "G"). "I" are
not supported, replace them with "N" instead.}

\item{subject_filepath}{A file path to a fasta or fastq file containing one
or multiple DNA sequences in which the pattern is searched.}

\item{subject_format}{A string, either "fasta" or "fastq".}

\item{max.mismatch}{The maximum number of mismatching letters
(maximum edit distance) allowed as integer.}

\item{with.indels}{A boolean, if TRUE then indels are allowed.}
}
\value{
A tibble containing a summary of the primer-template matches found.
}
\description{
Summarize primer-template matches
}
\examples{

# prepare pattern and subject
EV_fwd <- "GCTGCGYTGGCGGCC"

EV_sequences_fasta <- system.file("extdata", "Enterovirus_12059.fasta", package="ptmismatch")

summarize_matches(pattern = EV_fwd, subject_filepath = EV_sequences_fasta,
                  subject_format = "fasta", max.mismatch = 1, with.indels = TRUE)

}
